Update pal from commit: 47f44f10
* Build-time Settings Generation
* Add full-range copy support for EQAA image
* Cleanup and fixes Gfx9 APUs having ViewInstancing ExecuteIndirect failures.
* Add isShaderWriteable to BuildImageViewInfo
* Add flag lowZplanePolyOffsetBits
* Optimize BarrierAcquire() to group all non-ranged changed acquire sync into global sync
* Fix AcePreambleStream when compute engine is disabled
* Update devdriver to v22.04.19
* PS_DONE doesn't guarantee VS_DONE if no PS wave
* Fmask saturate error in RsrcProcMgr::ClearFmask
* Fix typo in RsrcPropMgr: pitchInMetaBlks should be mask down to 7 bits and sliceBits should be mask down to 6 bit
* Make optimization of clearing csBltActive to false more aggressively
* Reset gfxBltExecEopFenceVal/csBltExecEopFenceVal in GfxCmdBuffer::ResetState()
* Optimize GcrCntl only case in IssueReleaseSync()
* Remove command chunk's CPU reference counter
* Align SpillTable to cacheline size to prevent IB2 from reading potentially wrong data
* Fix ACE Offload Flag in ChipPropertes
* Validate Pipeline depth clamp
* Fix uint8 cast for RMT slice count
* Add Util::Abi::GfxIpVersionToMachineType
* Initial support for spoof any asic
* Add fixes for ExecuteIndirect packet
* Rework CmdAllocator's Trim to use a temp hash map
* Set proper stencilLayout in CopyDepthStencilImageGraphics
* Fix bug in MAX_PRIMS_PER_BATCH
* Fix CopyDepthStencilImageGraphics and ScaledCopyImageGraphics
* Fix the corruption observed on Door
* Add HIP API type enum for GpaSession
* Remove useAcqRelInterface flag
* AMF PugetBench performance enhancement through VCN decode power profile
* Add a "bool freeMemory" parameter to ICmdAllocator::Reset
* Add support for KMD Framelock
* Add settings service support
* Allow clients to set specific guardband clip ratios

+++ src/core/eventProvider.cpp
@@ -689,7 +689,7 @@ void EventProvider::LogResourceCreateEvent(
         imgCreateInfo.dimensions.dimension_Z = static_cast<uint16>(pImageData->pCreateInfo->extent.depth);
         imgCreateInfo.format = PalToRmtImageFormat(pImageData->pCreateInfo->swizzledFormat);
         imgCreateInfo.mips = static_cast<uint8>(pImageData->pCreateInfo->mipLevels);
-        imgCreateInfo.slices = static_cast<uint8>(pImageData->pCreateInfo->arraySize);
+        imgCreateInfo.slices = static_cast<uint16>(pImageData->pCreateInfo->arraySize);
         imgCreateInfo.samples = static_cast<uint8>(pImageData->pCreateInfo->samples);
         imgCreateInfo.fragments = static_cast<uint8>(pImageData->pCreateInfo->fragments);
         imgCreateInfo.tilingType = PalToRmtTilingType(pImageData->pCreateInfo->tiling);
